requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-dbsettings==1.3.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
